Add comment and use float instead of double

+++ examples/raytracing/example_raytracing.cpp
@@ -293,7 +293,7 @@ int main(int argc, char *argv[])
                   Kokkos::rand<GeneratorType, float>::draw(g, dz)};
 
           float upsilon = Kokkos::rand<GeneratorType, float>::draw(
-              g, 2.f * Kokkos::numbers::pi_v<double>);
+              g, 2.f * Kokkos::numbers::pi_v<float>);
           float theta =
               acos(1 - 2 * Kokkos::rand<GeneratorType, float>::draw(g));
           ArborX::Experimental::Vector direction{
